requirements.txt
pandas-profiling
folium